requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
dp==1.0
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
yarl==1.8.1
